moves kubeconfig creation logic inside svc account service
