+++ internal/resources/resources.go
@@ -2,7 +2,6 @@ package resources
 
 import (
 	"context"
-
 	k8sclient "k8s.io/client-go/kubernetes"
 )
